Fix CCK namespace and paths
