+++ compatibility/cck_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative 'support/shared_examples'
-require_relative 'support/compatibility_kit'
+require_relative 'support/cucumber/compatibility_kit'
 
 require 'cucumber/compatibility_kit'
