+++ compatibility/spec/cucumber/compatibility_kit_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require_relative '../support/compatibility_kit'
+require_relative '../../support/cucumber/compatibility_kit'
 
-RSpec.describe CompatibilityKit do
+RSpec.describe Cucumber::CompatibilityKit do
   let(:features_path) { File.expand_path("#{File.dirname(__FILE__)}/../features") }
 
   describe '.supporting_code_for' do
