+++ compatibility/support/cucumber/compatibility_kit/helpers.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+module Cucumber
+  module CompatibilityKit
+    module Helpers
+      def message_type(message)
+        message.to_h.each do |key, value|
+          return key unless value.nil?
+        end
+      end
+
+      def parse_ndjson_file(path)
+        parse_ndjson(File.read(path))
+      end
+
+      def parse_ndjson(ndjson)
+        Cucumber::Messages::Helpers::NdjsonToMessageEnumerator.new(ndjson)
+      end
+    end
+  end
+end
